Move FHIR server specific functionality into phdi_building_blocks.fhir.
