+++ FHIR-parquet-explore/read_write_demo.py
@@ -9,7 +9,11 @@ import pyarrow.parquet as pq
 import logging
 import numpy as np
 import random
-from phdi_building_blocks.fhir import AzureFhirserverCredentialManager
+from phdi_building_blocks.fhir import (
+    AzureFhirserverCredentialManager,
+    query_fhir_server,
+    log_fhir_server_error,
+)
 import os
 
 
@@ -25,36 +29,6 @@ def load_schema(path: str) -> dict:
     with open(path, "r") as file:
         schema = yaml.safe_load(file)
     return schema
-
-
-def query_fhir_server(
-    credential_manager: AzureFhirserverCredentialManager,
-    query: str = "",
-    specific_url: str = "",
-) -> requests.models.Response:
-    """
-    Given the url for a FHIR server, a query to execute via the server's API, and an
-    authentication method execute a query of the server for all resources of the
-    specified type and return the response.
-
-    :param str base_url: Url of the FHIR server to be queried.
-    :param AzureFhirserverCredentialManager credential_manager: A credential manager for
-    a FHIR server.
-    :param str query: The query for the FHIR server to execute.
-    :return requests.models.Response response: The response from the FHIR server.
-    """
-    # When no specific url is provided use the base url for the FHIR server from the
-    # credential manager and add the query. Otherwise use the specific url by itself.
-    if specific_url == "":
-        full_url = credential_manager.fhir_url + query
-    else:
-        full_url = specific_url + query
-
-    access_token = credential_manager.get_access_token().token
-    header = {"Authorization": f"Bearer {access_token}"}
-    response = requests.get(url=full_url, headers=header)
-
-    return response
 
 
 def parse_value(
@@ -174,20 +148,6 @@ def apply_schema_to_resource(resource: dict, schema: dict) -> dict:
             data[resource_schema[field]["new_name"]] = value
 
     return data
-
-
-def log_fhir_server_error(status_code: int):
-    """Given a status code from a FHIR server's response log the specified error.
-
-    :param int status_code: Status code returned by a FHIR server
-    """
-    if status_code == 401:
-        logging.error(
-            "FHIR SERVER ERROR - Status Code 401: Failed to authenticate with the FHIR server."
-        )
-
-    elif status_code == 404:
-        logging.error("FHIR SERVER ERROR - Status Code 404: FHIR server not found.")
 
 
 def make_resource_type_table(
